pyproject.toml
[project]
name = "promocode-service"
version = "0.1.0"
authors = [{ name = "Alex Pirogov", email = "pirogovalexandr00@mail.ru" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.4",
    "django-filter>=25.2",
    "djangorestframework>=3.17.1",
]

[dependency-groups]
dev = [
    "ruff>=0.15.9",
]

[tool.ruff]
line-length = 99

[tool.ruff.lint]
select = ["C", "E", "F", "W", "B"]
ignore = ["F541"]

[tool.ruff.lint.isort]
known-first-party = ["qasb_kit"]

[tool.ruff.format]
quote-style = "single"
